+++ Wire-iOS/Sources/UserInterface/Settings/CellDescriptors/SettingsPropertySelectValueCellDescriptor.swift
@@ -16,7 +16,6 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 // 
 
-
 import Foundation
 import WireSystem
 import UIKit
@@ -28,15 +27,15 @@ final class SettingsPropertySelectValueCellDescriptor: SettingsPropertyCellDescr
     let value: SettingsPropertyValue
     let title: String
     let identifier: String?
-    
-    typealias SelectActionType = (SettingsPropertySelectValueCellDescriptor) -> ()
+
+    typealias SelectActionType = (SettingsPropertySelectValueCellDescriptor) -> Void
     let selectAction: SelectActionType?
     let backgroundColor: UIColor?
     var visible: Bool = true
 
     weak var group: SettingsGroupCellDescriptorType?
     var settingsProperty: SettingsProperty
-    
+
     init(settingsProperty: SettingsProperty, value: SettingsPropertyValue, title: String, identifier: String? = .none, selectAction: SelectActionType? = .none, backgroundColor: UIColor? = .none) {
         self.settingsProperty = settingsProperty
         self.value = value
@@ -45,7 +44,7 @@ final class SettingsPropertySelectValueCellDescriptor: SettingsPropertyCellDescr
         self.selectAction = selectAction
         self.backgroundColor = backgroundColor
     }
-    
+
     func featureCell(_ cell: SettingsCellType) {
         cell.titleText = self.title
         cell.cellColor = self.backgroundColor
@@ -53,14 +52,14 @@ final class SettingsPropertySelectValueCellDescriptor: SettingsPropertyCellDescr
             valueCell.accessoryType = self.settingsProperty.value() == self.value ? .checkmark : .none
         }
     }
-    
+
     func select(_ value: SettingsPropertyValue?) {
         do {
             try settingsProperty.set(newValue: self.value)
         } catch (let e) {
             zmLog.error("Cannot set property: \(e)")
         }
-        
+
         selectAction?(self)
     }
 }
