+++ Wire-iOS/Sources/UserInterface/Settings/DarkThemeOption.swift
@@ -1,0 +1,59 @@
+
+// Wire
+// Copyright (C) 2020 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import Foundation
+
+enum DarkThemeOption: Int {
+    case dark, light, auto
+
+    static var settingKey: SettingKey = .twitterOpeningRawValue
+    static var defaultPreference: DarkThemeOption = .auto
+
+    var keyValueString: String {
+        switch self {
+        case .dark: return "dark"
+        case .light: return "light"
+        case .auto: return "auto"
+        }
+    }
+
+    var displayString: String {
+        switch self {
+        case .dark: return "dark_theme.option.dark".localized
+        case .light: return "dark_theme.option.light".localized
+        case .auto: return "dark_theme.option.auto".localized
+        }
+    }
+
+    static var allOptions: [DarkThemeOption] {
+        return [.dark, .light, .auto]
+    }
+
+    init?(keyValueString: String) {
+        switch keyValueString {
+        case "dark":
+            self = .dark
+        case "light":
+            self = .light
+        case "auto":
+            self = .auto
+        default:
+            return nil
+        }
+    }
+}
