+++ Wire-iOS/Sources/UserInterface/Settings/LinkOpening/TweetOpening.swift
@@ -21,47 +21,6 @@ import WireSystem
 
 private let log = ZMSLog(tag: "link opening")
 
-enum DarkThemeOption: Int {
-    case dark, light, auto
-
-    static var settingKey: SettingKey = .twitterOpeningRawValue
-    static var defaultPreference: DarkThemeOption = .auto
-
-    var keyValueString: String {
-        switch self {
-        case .dark: return "dark"
-        case .light: return "light"
-        case .auto: return "auto"
-        }
-    }
-    
-    var displayString: String {
-        switch self {
-        case .dark: return "dark_theme.option.dark".localized
-        case .light: return "dark_theme.option.light".localized
-        case .auto: return "dark_theme.option.auto".localized
-        }
-    }
-
-    static var allOptions: [DarkThemeOption] {
-        return [.dark, .light, .auto]
-    }
-    
-    init?(keyValueString: String) {
-        switch keyValueString {
-        case "dark":
-            self = .dark
-        case "light":
-            self = .light
-        case "auto":
-            self = .auto
-        default:
-            return nil
-        }
-    }
-}
-
-
 enum TweetOpeningOption: Int, LinkOpeningOption {
     case none, tweetbot, twitterrific
 
@@ -90,7 +49,6 @@ enum TweetOpeningOption: Int, LinkOpeningOption {
     }
 }
 
-
 extension URL {
 
     func openAsTweet() -> Bool {
@@ -99,7 +57,7 @@ extension URL {
         let saved = TweetOpeningOption.storedPreference
         log.debug("Saved option to open a tweet: \(saved.displayString)")
         let app = UIApplication.shared
-        
+
         switch saved {
         case .none: return false
         case .tweetbot:
@@ -111,15 +69,13 @@ extension URL {
             log.debug("Trying to open twitterific app using \"\(url)\"")
             app.open(url)
         }
-        
+
         return true
     }
 
 }
 
-
 // MARK: - Private
-
 
 fileprivate extension UIApplication {
 
@@ -133,7 +89,6 @@ fileprivate extension UIApplication {
 
 }
 
-
 extension URL {
 
     var isTweet: Bool {
@@ -141,7 +96,6 @@ extension URL {
     }
 
 }
-
 
 fileprivate extension URL {
 
@@ -173,11 +127,10 @@ fileprivate extension URL {
 
 }
 
-
 private extension String {
 
     func replacingWithTweetbotURLScheme(_ string: String) -> String {
         return replacingOccurrences(of: string, with: "tweetbot://")
     }
-    
+
 }
